% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{betas_to_consensus_vector}
\alias{betas_to_consensus_vector}
\title{Converts betas matrix to consensus vectors of probes for each group 
{VALIDATED}}
\usage{
betas_to_consensus_vector(
  reference_file,
  series_matrix_file,
  betas,
  group_names,
  remove_names
)
}
\arguments{
\item{reference_file}{directory location of where given IDAT file is located}

\item{series_matrix_file}{directory location of where given series matrix 
file is located}

\item{betas}{matrix (rows are cell names, columns are probe names)}

\item{group_names}{list of cell names in betas}

\item{remove_names}{list of cell names in betas}
}
\value{
consensus vector data frame for each cell in group names
}
\description{
This function creates a consensus vector for each group using the betas 
matrix
}
\examples{
consensus_vector <- betas_to_consensus_vector('~/references/
EPIC.hg38.manifest.rds', 'GSE110554/samples_GEOLoadSeriesMatrix.rds', 
betas, c('NK', "Monocytes", ...), c('A.mix', 'B.mix', ...))
... some visualization 
}
