% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{build_icceTree}
\alias{build_icceTree}
\title{Fully generate probe-node matrix with available data}
\usage{
build_icceTree(consensus_vector, tree)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}

\item{tree}{ape::phylo tree}
}
\value{
prone-node matrix storing the methylation status across all probes 
and nodes on a tree
}
\description{
This function fully generates the probe-node matrix with a given 
tree. It performs ancestral tree reconstruction using Fitch's algorithm.
}
\examples{
icceTree <- build_icceTree(consensus_vector, tree)
tree <- icceTree$tree
probe_node_matrix <- icceTree$probe_node_matrix
}
