% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{create_node_matrix}
\alias{create_node_matrix}
\title{Create node matrix used for tree walking functions}
\usage{
create_node_matrix(tree)
}
\arguments{
\item{tree}{ape::phylo tree}
}
\value{
node matrix with one column for ancestors and another column for 
descendants
}
\description{
Returns a matrix which tracks where recursive functions are in a tree 
structure. This is passed to most recursive functions in this package.
}
\examples{
node_matrix <- create_node_matrix(tree)
}
