% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{cut_tree}
\alias{cut_tree}
\title{Recursively walk through tree and remove descendants.}
\usage{
cut_tree(node, root, node_matrix, i)
}
\arguments{
\item{node}{node number}

\item{root}{root of non-proper subtree}

\item{node_matrix}{matrix obtained from create_node_matrix function}

\item{i}{row in node matrix, start at 1}
}
\value{
None
}
\description{
This function will recursively walk through a given tree and remove all of 
node n's descendants, making node n a new leaf.
}
\examples{
cut_tree(node, get_root(tree), node_matrix, 1)
... some visualization 
}
