% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{fit_betas_to_distribution}
\alias{fit_betas_to_distribution}
\title{Categorize betas based on bimodal distribution {VALIDATED}}
\usage{
fit_betas_to_distribution(betas, bi)
}
\arguments{
\item{betas}{matrix (rows are cell names, columns are probe names)}

\item{bi}{matrix specifying mu0, mu1, sigma0, sigma1 across all cell types}
}
\value{
betas categorized based on bimodal distribution
}
\description{
This function uses a given bimodal distribution generated within cells
in order to categorize betas as either methylated (1), unmethylated (0), or 
ambiguous (-1).
}
\examples{
betas <- fit_betas_to_distribution(betas, bi)
... some visualization 
}
