% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{fit_betas_to_distribution_p}
\alias{fit_betas_to_distribution_p}
\title{Categorize betas based on bimodal distribution and return the p-value of 
each categorization.}
\usage{
fit_betas_to_distribution_p(betas, bi)
}
\arguments{
\item{betas}{matrix (rows are cell names, columns are probe names)}

\item{bi}{matrix specifying mu0, mu1, sigma0, sigma1 across all cell types}
}
\value{
betas matrix categorized based on bimodal distribution
}
\description{
This function uses a given bimodal distribution generated within cell types
in order to categorize betas as either methylated (1) or unmethylated (0) 
in the betas structure. It also returns the p-value of each categorization 
in the betas_p structure.
}
\examples{
betas <- fit_betas_to_distribution(betas, bi)
}
