% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{generate_bimodal_distribution}
\alias{generate_bimodal_distribution}
\title{Generate bimodal distribution {VALIDATED}}
\usage{
generate_bimodal_distribution(betas)
}
\arguments{
\item{betas}{matrix (rows are cell names, columns are probe names)}
}
\value{
matrix specifying mu0, mu1, sigma0, sigma1 across all cell types
}
\description{
This function generates a bimodal distribution on the betas matrix data 
within cells
}
\examples{
bi <- generate_bimodal_distribution(betas)
... some visualization 
}
