% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{generate_p_values}
\alias{generate_p_values}
\title{Work on this...}
\usage{
generate_p_values(x, mu0, mu1, nsamples = 10)
}
\arguments{
\item{x}{data structure}

\item{mu0}{name of variable/file}

\item{mu1}{directory to which the data should be saved}

\item{nsamples}{number of cell types}
}
\value{
None
}
\description{
Given...
}
\examples{
save_data(betas, 'betas', getwd())
}
