% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{generate_probe_node_matrix}
\alias{generate_probe_node_matrix}
\title{Fully generate probe-node matrix with available data {VALIDATED}}
\usage{
generate_probe_node_matrix(consensus_vector, tree)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}

\item{tree}{ape::phylo tree}
}
\value{
probe_node_matrix
}
\description{
This function fully generates the probe-node matrix with a given 
tree
}
\examples{
tree_info <- generate_probe_node_matrix(consensus_vector, 
upgma_tree(consensus_vector))
tree <- tree_info[1]
probe_node_matrix <- tree_info[2]
... some visualization 
}
