% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{get_leaves}
\alias{get_leaves}
\title{Leaves of a tree}
\usage{
get_leaves(tree)
}
\arguments{
\item{tree}{ape::phylo tree}
}
\value{
leafs on a given tree
}
\description{
Given a tree, this function returns the leaves
}
\examples{
leaves <- get_leaves(tree)
... some visualization 
}
