% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{get_probes_of_gene}
\alias{get_probes_of_gene}
\title{Probes along a given gene {VALIDATED}}
\usage{
get_probes_of_gene(reference, gene)
}
\arguments{
\item{reference}{reference file}

\item{gene}{gene name}
}
\value{
list of probes along a particular gene
}
\description{
This function returns a list of probes along a particular gene
}
\examples{
gene_probes <- get_probes_of_gene(reference, gene)
... some visualization 
}
