% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{get_reference}
\alias{get_reference}
\title{Returns reference file {VALIDATED}}
\usage{
get_reference(reference_file)
}
\arguments{
\item{reference_file}{directory location of where given IDAT file is located}
}
\value{
beta matrix without probes located on sex chromosomes.
}
\description{
This function returns a reference file given location. Any preprocessing 
steps for the reference will also be added in this function.
}
\examples{
reference <- get_reference('~/references/hg38/annotation/R/
EPIC.hg38.manifest.rds')
... some visualization 
}
