% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{get_root}
\alias{get_root}
\title{Root of a tree {VALIDATED}}
\usage{
get_root(tree)
}
\arguments{
\item{tree}{ape::phylo tree}
}
\value{
node corresponding the the root of the tree
}
\description{
Returns the root of a tree structure.
}
\examples{
root <- get_root(tree)
... some visualization 
}
