% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{grow_off_leaf}
\alias{grow_off_leaf}
\title{Grow a tree at a given leaf node}
\usage{
grow_off_leaf(tree, leaf, node_names)
}
\arguments{
\item{tree}{ape::phylo tree}

\item{leaf}{node number}

\item{node_names}{list of two node names}
}
\value{
new tree topology--two branches added to given leaf and labeled 
accordingly
}
\description{
Add a pair of leaves branching off of the given leaf node for the tree. 
Asigns the two newly added leaves labels from node_names the old leaf where 
the two branches were add as an internal node.
}
\examples{
tree <- grow_off_leaf(tree, 1, c('Fibroblasts', 'Fibrocytes'))
... some visualization 
}
