% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icceTree.R
\name{icce_tree}
\alias{icce_tree}
\title{Create icce tree object}
\usage{
icce_tree(probe_node_matrix, probe_node_matrix_p, tree)
}
\arguments{
\item{probe_node_matrix}{probe-node matrix storing the methylation status 
across all probes and nodes on a tree}

\item{probe_node_matrix_p}{probe-node matrix storing the p-value of each 
node across all probes on a tree}

\item{tree}{ape::phylo tree}
}
\value{
icce tree object
}
\description{
Returns icce tree object given a probe-node matrix, a probe-node p-values 
and a ape::phylo tree
}
\examples{
reconstruct_internal_nodes(get_root(tree), node_matrix, 1)
}
