% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{inherit_parental_state}
\alias{inherit_parental_state}
\title{Inherit parental states to ambiguous nodes}
\usage{
inherit_parental_state(root, node_matrix, i)
}
\arguments{
\item{root}{root of non-proper subtree}

\item{node_matrix}{matrix obtained from create_node_matrix function}

\item{i}{row in node matrix, start at 1}
}
\value{
None
}
\description{
This function will recursively walk through a given tree and assigns 
parental states to ambiguous children nodes
}
\examples{
inherit_parental_state(get_root(tree), node_matrix, 1)
}
