% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{initialize_probe_node_matrix}
\alias{initialize_probe_node_matrix}
\title{Initialize the probe-node matrix}
\usage{
initialize_probe_node_matrix(consensus_vector, tree)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}

\item{tree}{ape::phylo tree}
}
\value{
initialized probe_node_matrix
}
\description{
This function simply initializes a probe-node matrix across given nodes and 
probes with zeros.
}
\examples{
probe_node_matrix <- initialize_probe_node_matrix(consensus_vector, 
upgma_tree(consensus_vector))
}
