% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{ninternal_nodes}
\alias{ninternal_nodes}
\title{Number of internal nodes on a tree {VALIDATED}}
\usage{
ninternal_nodes(tree)
}
\arguments{
\item{tree}{ape::phylo tree}
}
\value{
number of internal nodes on a given tree
}
\description{
Given a tree, this function returns the number of internal nodes on tree.
}
\examples{
number_of_internal_nodes = ninternal_nodes(tree)
... some visualization 
}
