% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{nleaves}
\alias{nleaves}
\title{Get number of leaves of a tree}
\usage{
nleaves(icceTree)
}
\arguments{
\item{icceTree}{icceTree data structure}
}
\value{
number of leafs of a given tree
}
\description{
Returns the number of leaves on a given tree.
}
\examples{
number_of_leaves = nleaves(icceTree)
}
