% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{nnodes}
\alias{nnodes}
\title{Get total number of nodes on a tree}
\usage{
nnodes(icceTree)
}
\arguments{
\item{icceTree}{icceTree data structure}
}
\value{
number of nodes on a tree
}
\description{
Returns the number of nodes on a given tree.
}
\examples{
number_of_nodes = nnodes(icceTree)
}
