% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{nprobes}
\alias{nprobes}
\title{Get number of probes in consensus vector}
\usage{
nprobes(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
number of probes
}
\description{
Returns the number of probes in a list of consensus vectors
}
\examples{
number_of_probes <- nprobes(consensus_vector)
... some visualization 
}
