% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{read_IDAT}
\alias{read_IDAT}
\title{Load betas data}
\usage{
read_IDAT(IDAT_dir)
}
\arguments{
\item{IDAT_dir}{directory location of where given IDAT file is located}
}
\value{
betas matrix
}
\description{
Using sesame, this function loads betas data using IDAT files stored in a 
given directory
}
\examples{
betas <- read_IDAT('~/20191212_GEO_datasets/GSE110554')
}
