% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{reconstruct_internal_nodes}
\alias{reconstruct_internal_nodes}
\title{Reconstruct internal nodes}
\usage{
reconstruct_internal_nodes(root, node_matrix, i)
}
\arguments{
\item{root}{root of non-proper subtree}

\item{node_matrix}{matrix obtained from create_node_matrix function}

\item{i}{row in node matrix, start at 1}
}
\value{
None
}
\description{
This function will recursively walk through a given tree and reconstruct 
all of its internal states using Fitch's algorithm
}
\examples{
reconstruct_internal_nodes(get_root(tree), node_matrix, 1)
}
