% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{remove_branches_from_node}
\alias{remove_branches_from_node}
\title{Cut tree at specified node}
\usage{
remove_branches_from_node(tree, node)
}
\arguments{
\item{tree}{ape::phylo tree}

\item{node}{node number}
}
\value{
new tree topology--node is now a leaf and all of node's descendants 
are removed.
}
\description{
Given a tree and a selected node, this function removes the branches 
stemming off of that node, thus making that node a leaf. It returns this 
altered tree.
}
\examples{
tree <- remove_branches_from_node(tree, node)
... some visualization 
}
