% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{remove_cell_types}
\alias{remove_cell_types}
\title{Remove specific cell types from betas matrix {VALIDATED}}
\usage{
remove_cell_types(betas, remove_names)
}
\arguments{
\item{betas}{matrix (rows are cell names, columns are probe names)}

\item{remove_names}{list of cell names in betas}
}
\value{
beta matrix without specified cell names
}
\description{
This function removes specified group names. Common names include MIX or 
are typically heterogeneous mixtures of cells.
}
\examples{
betas <- remove_cell_types(betas, c('A.mix', 'B.mix', ...))
... some visualization 
}
