% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{remove_intermediate_probes}
\alias{remove_intermediate_probes}
\title{Removes intermediate probes}
\usage{
remove_intermediate_probes(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
consensus vector without intermediate probes
}
\description{
Returns consensus vector with intermediate probes removed.
}
\examples{
consensus_vector <- remove_intermediate_probes(consensus_vector)
... some visualization 
}
