% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{remove_NA_probes}
\alias{remove_NA_probes}
\title{Removes NA probes {VALIDATED}}
\usage{
remove_NA_probes(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
consensus vector without NA probes
}
\description{
Selects and removes NA probes from consensus vectors
}
\examples{
consensus_vector <- remove_NA_probes(consensus_vector)
... some visualization 
}
