% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{remove_sex_chr}
\alias{remove_sex_chr}
\title{Removes probes on sex chromosomes}
\usage{
remove_sex_chr(reference, betas)
}
\arguments{
\item{reference}{reference file}

\item{betas}{matrix (rows are probe names, columns are cell names)}
}
\value{
betas matrix without probes located on sex chromosomes.
}
\description{
This function removes probes located on sex chromosomes because these 
probes have a disproportionate global methylation status
}
\examples{
betas <- remove_sex_chr(reference, betas)
}
