% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{remove_unapplicable_probes}
\alias{remove_unapplicable_probes}
\title{Removes all inapplicable probes}
\usage{
remove_unapplicable_probes(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
consensus vector without inapplicable probes
}
\description{
Returns consensus vector with NA, intermediate, and zero parsimony probes 
removed
}
\examples{
consensus_vector <- remove_unapplicable_probes(consensus_vector)
}
