% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{remove_zero_parsimony_probes}
\alias{remove_zero_parsimony_probes}
\title{Removes zero parsimony probes}
\usage{
remove_zero_parsimony_probes(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
consensus vector without zero parsimony probes
}
\description{
Returns consensus vector with zero parsimony probes removed
}
\examples{
consensus_vector <- remove_zero_parsimony_probes(consensus_vector)
... some visualization 
}
