% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{save_data}
\alias{save_data}
\title{Saves specified data with variable name to working directory}
\usage{
save_data(data, variable_name, dir)
}
\arguments{
\item{data}{data structure}

\item{variable_name}{name of variable/file}

\item{dir}{directory to which the data should be saved}
}
\value{
None
}
\description{
Given data, a variable name, and a location for that data to be stored, 
this function simply saves data.
}
\examples{
save_data(betas, 'betas', getwd())
}
