% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{show_gene_track}
\alias{show_gene_track}
\title{Track the single probe resolution change of a gene}
\usage{
show_gene_track(
  filename,
  tree,
  gene,
  thermo_prop_internal_nodes,
  thermo_prop_leaves
)
}
\arguments{
\item{filename}{filename to which the plot will be saved}

\item{tree}{ape::phylo tree}

\item{gene}{gene name}

\item{thermo_prop_leaves}{thermo proportion plot for leaves of all of
the probe states}

\item{thermo_prop_node}{thermo proportion plot for internal nodes of all of 
the probe states}
}
\value{
None
}
\description{
Displays constructed tree with node labels for the proportion of each probe 
state for a selected gene. This is saved at the given filename.
}
\examples{
show_gene_track('tree_CD81', tree=icceTree$tree, gene='CD81', 
thermo_prop_internal_nodes=thermo_prop_internal_nodes, thermo_prop_leaves=
thermo_prop_leaves)
... some visualization 
}
