% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{show_heatmap}
\alias{show_heatmap}
\title{Display heatmap of probes {In progress}}
\usage{
show_heatmap(betas, reference)
}
\arguments{
\item{betas}{betas}

\item{reference}{reference}
}
\value{

}
\description{
This is in progress
}
\examples{
head_map <- show_heatmap()
}
