% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{track_gene}
\alias{track_gene}
\title{Proportion of probes methylated, unmethylated, ambiguous at each node 
across a tree {VALIDATED}}
\usage{
track_gene(tree, probe_node_matrix, reference, gene)
}
\arguments{
\item{tree}{ape::phylo tree}

\item{probe_node_matrix}{matrix storing the methylation status across all
probes and nodes on a tree}

\item{reference}{reference file}

\item{gene}{gene name}
}
\value{
a structure composed of two lists that contain methylation state 
proportions for internal nodes and methylation state proportions for leaves
}
\description{
This function returns the proportion of each probes that are methylated, 
unmethylated, ambiguous at each node across a tree.
}
\examples{
gene_information <- track_gene(tree, probe_node_matrix, reference, 'CD81')
thermo_prop_internal_nodes = gene_information$thermo_prop_internal_nodes
thermo_prop_leaves = gene_information$thermo_prop_leaves
... some visualization 
}
