% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{upgma_tree}
\alias{upgma_tree}
\title{Get UPGMA tree construction}
\usage{
upgma_tree(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
ape::phylo tree constructed using UPGMA algorithm
}
\description{
This function uses the UPGMA algorithm to construct a phylogenetic tree 
using methylation data.
}
\examples{
tree <- upgma_tree(consensus_vector)
}
