% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{add_data}
\alias{add_data}
\title{Add to probe-node matrix at a new node}
\usage{
add_data(probe_node_matrix, n_new., data)
}
\arguments{
\item{probe_node_matrix}{matrix storing the methylation status across all
probes and nodes on a tree}

\item{data}{column matrix of methylation status for a node across many 
probes}

\item{n_new}{node number}
}
\value{
probe-node matrix with data stored at node n_new
}
\description{
At a given node, this function adds data to the probe-node matrix at node 
n_new. Make sure that the sizes of data and the probe-node matrix are 
compatible before calling this function.
}
\examples{
tree <- add_data(probe_node_matrix, 12, probe_node_matrix['6'])
... some visualization 
}
