% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{create_node_matrix}
\alias{create_node_matrix}
\title{Node matrix used for tree walking {VALIDATED}}
\usage{
create_node_matrix(tree)
}
\arguments{
\item{tree}{ape::phylo tree}
}
\value{
node matrix with one column for ancestors and another for 
descendants
}
\description{
Returns a matrix which tracks where recursive functions are in a tree 
structure. This is past to most recursive functions in the pipeline.
}
\examples{
node_matrix <- create_node_matrix(tree)
... some visualization 
}
