% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{cut_tree}
\alias{cut_tree}
\title{Recursively walk through a tree and structure remove descendants of given 
node.}
\usage{
cut_tree(node, root, node_matrix, i)
}
\arguments{
\item{node}{node number}

\item{root}{root of non-proper subtree}

\item{node_matrix}{node matrix obtained from create_node_matrix function}

\item{i}{row in node matrix, starting at 1}
}
\value{
None
}
\description{
This function will recursively walk through a given tree and remove all of 
the descendants of a given node, making the given node a new leaf.
}
\examples{
cut_tree(node, get_root(tree), node_matrix, 1)
}
