% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{get_leaves}
\alias{get_leaves}
\title{Get the leaves of a tree}
\usage{
get_leaves(icceTree)
}
\arguments{
\item{icceTree}{icceTree data structure}
}
\value{
a list of nodes representing the leaves of the tree
}
\description{
Given a tree, this function returns its leaves.
}
\examples{
leaves <- get_leaves(icceTree)
}
