% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{get_reference}
\alias{get_reference}
\title{Return reference data frame}
\usage{
get_reference(reference_file)
}
\arguments{
\item{reference_file}{file location of reference file}
}
\value{
reference file
}
\description{
This function returns the contents of a given reference file as a data 
frame. The rows of the data frame are the probe names, allowing for quick
indexing.
}
\examples{
reference <- get_reference('~/references/hg38/annotation/R/
EPIC.hg38.manifest.rds')
}
