% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{get_root}
\alias{get_root}
\title{Get root of a tree}
\usage{
get_root(icceTree)
}
\arguments{
\item{icceTree}{icceTree data structure}
}
\value{
node corresponding the the root of the tree
}
\description{
Returns the root of a given tree structure.
}
\examples{
root <- get_root(icceTree)
}
