% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{grow_off_leaf}
\alias{grow_off_leaf}
\title{Grow a tree at a given leaf node}
\usage{
grow_off_leaf(icceTree, leaf, node_names)
}
\arguments{
\item{icceTree}{icceTree data structure}

\item{leaf}{node number}

\item{node_names}{list of two node names}
}
\value{
new tree topology--two branches added to given leaf and labeled 
accordingly
}
\description{
Add a pair of leaves branching off of a given leaf node on a tree. 
Assigns the two newly added leaves labels from node_names. The old leaf 
where the two new branches were add is also assigned as an internal node.
}
\examples{
icceTree <- grow_off_leaf(icceTree, 1, c('Fibroblasts', 'Fibrocytes'))
}
