% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{ninternal_nodes}
\alias{ninternal_nodes}
\title{Get number of internal nodes on a tree}
\usage{
ninternal_nodes(icceTree)
}
\arguments{
\item{icceTree}{icceTree data structure}
}
\value{
number of internal nodes on a given tree
}
\description{
Returns the number of internal nodes on the given tree.
}
\examples{
number_of_internal_nodes = ninternal_nodes(icceTree)
}
