% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{nleaves}
\alias{nleaves}
\title{Number of leaves of a tree}
\usage{
nleaves(tree)
}
\arguments{
\item{tree}{ape::phylo tree}
}
\value{
number of leafs of a given tree
}
\description{
Given a tree, this function returns the number of leaves on a tree.
}
\examples{
number_of_leaves = nleaves(tree)
... some visualization 
}
