% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{nnodes}
\alias{nnodes}
\title{Total number of nodes on a tree {VALIDATED}}
\usage{
nnodes(tree)
}
\arguments{
\item{tree}{ape::phylo tree}
}
\value{
number of nodes on a tree
}
\description{
Given a tree, this function returns the number of nodes on a tree.
}
\examples{
number_of_nodes = nnodes(tree)
... some visualization 
}
