% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTree.R
\name{nprobes}
\alias{nprobes}
\title{Number of probes in data {VALIDATED}}
\usage{
nprobes(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
number of probes
}
\description{
This function returns the number of probes in the data
}
\examples{
number_of_probes <- nprobes(consensus_vector)
... some visualization 
}
