% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTree.R
\name{relabel_tree}
\alias{relabel_tree}
\title{Relabel node on tree}
\usage{
relabel_tree(tree, n_old, n_new)
}
\arguments{
\item{tree}{ape::phylo tree}

\item{n_old}{node number}

\item{n_new}{node number}
}
\value{
relabeled tree with node label n_old with node label
n_new
}
\description{
Relabel node label n_old with node label n_new on tree.
}
\examples{
tree <- relabel_tree(tree, 7, 10)
... some visualization 
}
