% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessData.R
\name{remove_unapplicable_probes}
\alias{remove_unapplicable_probes}
\title{Removes all inapplicable probes {VALIDATED}}
\usage{
remove_unapplicable_probes(consensus_vector)
}
\arguments{
\item{consensus_vector}{consensus vector data frame for each cell in group 
names}
}
\value{
consensus vector without inapplicable probes
}
\description{
Selects and removes NA, intermediate, and zero parsimony probes from 
consensus vectors
}
\examples{
consensus_vector <- remove_unapplicable_probes(consensus_vector)
... some visualization 
}
