% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{show_heatmap}
\alias{show_heatmap}
\title{Relabel node in probe-node matrix}
\usage{
show_heatmap()
}
\arguments{
\item{probe_node_matrix}{matrix storing the methylation status across all
probes and nodes on a tree}

\item{n_old}{node number}

\item{n_new}{node number}
}
\value{
relabeled probe_node_matrix with node label n_old with node label
n_new
}
\description{
Relabel node label n_old with node label n_new in probe-node matrix.
}
\examples{
probe_node_matrix <- relabel_node_matrix(probe_node_matrix, 7, 10)
... some visualization 
}
