% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{show_tree}
\alias{show_tree}
\title{Display tree}
\usage{
show_tree(filename, tree, title, edge_labels)
}
\arguments{
\item{filename}{filename to which the plot will be saved}

\item{tree}{ape::phylo tree}

\item{title}{title name}

\item{edge_labels}{optional labels for each branch}
}
\value{
None
}
\description{
Displays constructed tree and saves it at the given filename.
}
\examples{
show_tree(filename='tree.pdf', tree=icceTree$tree, title='Hematopoietic Tree', 
edge_labels=meth_gains_label) 
}
